refactor(llm): move stream content parser to the adapter module
Relocate _make_stream_content_parser / _make_think_tag_splitter into llm_adapter
so adapters can reuse them without a chat↔adapter import cycle.
